Azure/WhitelistIpAddressForPowerPlatform/requirements.txt
azure-mgmt-resource
azure-identity
azure-keyvault
azure-mgmt-keyvault